refactor: replace services by collector
