+++ plugins/all/all.go
@@ -6,23 +6,23 @@ import (
 	_ "github.com/hidracloud/hidra/v3/plugins/misc/string"
 
 	// Run http initialization
-	_ "github.com/hidracloud/hidra/v3/plugins/services/http"
+	_ "github.com/hidracloud/hidra/v3/plugins/collector/http"
 
 	// Run http initialization
-	_ "github.com/hidracloud/hidra/v3/plugins/services/dns"
+	_ "github.com/hidracloud/hidra/v3/plugins/collector/dns"
 
 	// Run ftp initialization
-	_ "github.com/hidracloud/hidra/v3/plugins/services/ftp"
+	_ "github.com/hidracloud/hidra/v3/plugins/collector/ftp"
 
 	// Run icmp initialization
-	_ "github.com/hidracloud/hidra/v3/plugins/services/icmp"
+	_ "github.com/hidracloud/hidra/v3/plugins/collector/icmp"
 
 	// Run tcp initialization
-	_ "github.com/hidracloud/hidra/v3/plugins/services/tcp"
+	_ "github.com/hidracloud/hidra/v3/plugins/collector/tcp"
 
 	// Run udp initialization
-	_ "github.com/hidracloud/hidra/v3/plugins/services/udp"
+	_ "github.com/hidracloud/hidra/v3/plugins/collector/udp"
 
 	// Run udp initialization
-	_ "github.com/hidracloud/hidra/v3/plugins/services/tls"
+	_ "github.com/hidracloud/hidra/v3/plugins/collector/tls"
 )
