+++ plugins/collector/browser/browser.go
